moved Timer to tools.py
