+++ ipi/pes/extmace.py
@@ -34,8 +34,6 @@ class Extended_MACE_driver(MACE_driver):
 
 
 # --------------------------------------- #
-import time
-from datetime import datetime
 import torch
 import numpy as np
 from typing import List, Dict
@@ -43,6 +41,8 @@ from mace.tools.torch_geometric.batch import Batch
 from mace.calculators import MACECalculator
 from mace.modules.utils import get_outputs
 from ase.calculators.calculator import Calculator, all_changes
+
+from .tools import Timer
 
 
 class Extended_MACECalculator(MACECalculator):
@@ -437,37 +437,3 @@ def compute_dielectric_gradients(
         del gradient  # cleanup
     del grad_outputs  # cleanup
     return torch.stack(d_dielectric_dr, dim=0)  # [Pxyz,atoms,Rxyz]
-
-
-# --------------------------------------- #
-class Timer:
-    _stack = []
-    _records = []
-
-    def __init__(self, name, log=True):
-        self.name = name
-        self.log = log
-        self.level = len(Timer._stack)
-
-    def __enter__(self):
-        self.start = time.time()
-        Timer._stack.append(self)
-        return self
-
-    def __exit__(self, exc_type, exc_val, exc_tb):
-        self.end = time.time()
-        self.elapsed = self.end - self.start
-        Timer._stack.pop()
-        if self.log:
-            Timer._records.append((self.start, self.level, self.name, self.elapsed))
-
-    @staticmethod
-    def report():
-        for t, level, name, elapsed in sorted(Timer._records):
-            timestamp = datetime.fromtimestamp(t).strftime("%Y-%m-%d %H:%M:%S")
-            indent = "    " * level
-            print(f"{timestamp}{indent} {name}: {elapsed:.4f} s")
-
-    @staticmethod
-    def reset():
-        Timer._records.clear()
